:hammer: Refactor move & rename
