--- app/src/main/java/network/omisego/omgwallet/data/contract/BalanceDataRepository.kt
@@ -1,16 +1,0 @@
-package network.omisego.omgwallet.data.contract
-
-/*
- * OmiseGO
- *
- * Created by Phuchit Sirimongkolsathien on 22/9/2018 AD.
- * Copyright © 2017-2018 OmiseGO. All rights reserved.
- */
-
-import androidx.lifecycle.MutableLiveData
-import network.omisego.omgwallet.livedata.Event
-import network.omisego.omgwallet.model.APIResult
-
-interface BalanceDataRepository {
-    fun loadWallet(liveAPIResult: MutableLiveData<Event<APIResult>>)
-}
