+++ app/src/main/java/network/omisego/omgwallet/network/ClientConfig.kt
@@ -6,7 +6,7 @@ package network.omisego.omgwallet.network
  * Created by Phuchit Sirimongkolsathien on 5/10/2018 AD.
  * Copyright © 2017-2018 OmiseGO. All rights reserved.
  */
-interface ClientSetup {
+interface ClientConfig {
     val baseURL: String
     val socketBaseURL: String
     val apiKey: String
