+++ app/src/main/java/network/omisego/omgwallet/screen/auth/profile/main/ProfileViewModel.kt
@@ -13,9 +13,8 @@ import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.MutableLiveData
 import co.infinum.goldfinger.Goldfinger
 import network.omisego.omgwallet.BuildConfig
-import network.omisego.omgwallet.data.LocalRepository
-import network.omisego.omgwallet.livedata.Event
-import network.omisego.omgwallet.network.ClientProvider
+import network.omisego.omgwallet.repository.LocalRepository
+import network.omisego.omgwallet.util.Event
 import network.omisego.omgwallet.state.FingerprintDialogState
 
 class ProfileViewModel(
@@ -49,7 +48,7 @@ class ProfileViewModel(
 
     init {
         liveVersionName.value = BuildConfig.VERSION_NAME
-        liveEndpoint.value = ClientProvider.clientSetup.baseURL
+        liveEndpoint.value = BuildConfig.CONFIG_BASE_URL
     }
 
     fun hasFingerprintSupport() = Goldfinger.Builder(app).build().hasFingerprintHardware()
