+++ app/src/main/java/network/omisego/omgwallet/util/Event.kt
@@ -1,4 +1,4 @@
-package network.omisego.omgwallet.livedata
+package network.omisego.omgwallet.util
 
 /*
  * OmiseGO
