+++ app/src/main/java/network/omisego/omgwallet/util/EventObserver.kt
@@ -1,4 +1,4 @@
-package network.omisego.omgwallet.livedata
+package network.omisego.omgwallet.util
 
 import androidx.lifecycle.Observer
